requirements.txt
appdirs==1.4.4
black==0.0.0
click==7.1.2
dataclasses==0.7
flake8==3.8.4
grpcio==1.33.2
grpcio-tools==1.33.2
importlib-metadata==2.0.0
isort==5.6.4
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.0
protobuf==3.13.0
pycodestyle==2.6.0
pyflakes==2.2.0
regex==2020.10.28
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
